+++ src/helpers/contact.ts
@@ -1,4 +1,6 @@
-import { ContactFunction, Gender } from '../clients/server.generated';
+import {
+  Contact, ContactFunction, ContactSummary, Gender,
+} from '../clients/server.generated';
 
 export function formatContactName(
   fName: string | undefined,
@@ -30,7 +32,9 @@ export function formatFunctionShort(func: ContactFunction) {
     case ContactFunction.PRIMARY: return 'Primary';
     case ContactFunction.FINANCIAL: return 'Financial';
     case ContactFunction.OLD: return 'Old';
-    default: return 'Unknown';
+    default: {
+      return 'Unknown';
+    }
   }
 }
 
@@ -42,4 +46,45 @@ export function formatGender(gender: Gender) {
     case Gender.UNKNOWN: return 'Unknown gender';
     default: return 'Unknown gender';
   }
+}
+
+export function sortContactsByFunction(
+  contacts: Contact[] | ContactSummary[],
+  sortAlphabetically?: boolean,
+): Contact[] | ContactSummary[] {
+  if (sortAlphabetically) {
+    contacts.sort((c1, c2) => {
+      const n1 = formatContactName(c1.firstName, c1.lastNamePreposition, c1.lastName).toUpperCase();
+      const n2 = formatContactName(c2.firstName, c2.lastNamePreposition, c2.lastName).toUpperCase();
+      if (n1 < n2) return -1;
+      if (n1 > n2) return 1;
+      return 0;
+    });
+  }
+
+  contacts.sort((c1, c2) => {
+    switch (c1.function) {
+      case ContactFunction.PRIMARY: return -1;
+      case ContactFunction.NORMAL:
+        switch (c2.function) {
+          case ContactFunction.PRIMARY: return 1;
+          case ContactFunction.NORMAL: return 0;
+          case ContactFunction.FINANCIAL: return -1;
+          case ContactFunction.OLD: return -1;
+          default: return 0;
+        }
+      case ContactFunction.FINANCIAL:
+        switch (c2.function) {
+          case ContactFunction.PRIMARY: return 1;
+          case ContactFunction.NORMAL: return 1;
+          case ContactFunction.FINANCIAL: return 0;
+          case ContactFunction.OLD: return -1;
+          default: return 0;
+        }
+      case ContactFunction.OLD: return 1;
+      default: return 0;
+    }
+  });
+
+  return contacts;
 }